### 2019-06-18 #315 @ShixiangWang
Submission: UCSCXenaTools
Submitting Author: ShixiangWang (@ShixiangWang)
Repository: https://github.com/ShixiangWang/UCSCXenaTools
Version submitted: 1.2.3
Version accepted: TBD

---

```
Package: UCSCXenaTools
Title: Download and Explore Datasets from UCSC Xena Data Hubs
Version: 1.2.3
Authors@R: c(person("Shixiang", "Wang", email = "w_shixiang@163.com", role = c("aut", "cre"),
            person("Martin", "Morgan", role="aut"))
Maintainer: Shixiang Wang <w_shixiang@163.com>
Description: Download and explore datasets from UCSC Xena data hubs, which are
    a collection of UCSC-hosted public databases such as TCGA, ICGC, TARGET, GTEx, CCLE, and others.
    Databases are normalized so they can be combined, linked, filtered, explored and downloaded.
    httr,
    shinydashboard,
    shiny,
    DT
    prettydoc,
    covr
```

The package falls under the **data retrieval category** because it pulls data from [UCSC Xena Hubs](https://xenabrowser.net/datapages/).

    -  Target Audience:  cancer and clinical researchers, bioinformaticians
    - Applications: genomic and clinical analyses

None that I am aware of. There is a Python package called [xenaPython](https://github.com/ucscXena/xenaPython) which is the official API for Xena Hubs. However, **UCSCXenaTools** is more powerful and provides more features including download whole dataset and query a small amount of data.

https://github.com/ropensci/software-review/issues/314

This package is already available at CRAN
