+++ README.md
@@ -1,25 +1,154 @@
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 
-# UCSCXenaTools <img src='man/figures/logo.png' align="right" height="140" width="120" alt="logo"/>
-
 <!-- badges: start -->
 
-[![CRAN](http://www.r-pkg.org/badges/version-last-release/UCSCXenaTools)](https://cran.r-project.org/package=UCSCXenaTools)
-[![AppVeyor Build
-Status](https://ci.appveyor.com/api/projects/status/github/ShixiangWang/UCSCXenaTools?branch=master&svg=true)](https://ci.appveyor.com/project/ShixiangWang/UCSCXenaTools)
-[![Travis build
-status](https://travis-ci.org/ShixiangWang/UCSCXenaTools.svg?branch=master)](https://travis-ci.org/ShixiangWang/UCSCXenaTools)
+<table class="table">
 
-![](http://cranlogs.r-pkg.org/badges/UCSCXenaTools)
-![](http://cranlogs.r-pkg.org/badges/grand-total/UCSCXenaTools)
-[![Coverage
-Status](https://img.shields.io/codecov/c/github/ShixiangWang/UCSCXenaTools/master.svg)](https://codecov.io/github/ShixiangWang/UCSCXenaTools?branch=master)
-[![GitHub
-issues](https://img.shields.io/github/issues/ShixiangWang/UCSCXenaTools.svg)](https://github.com/ShixiangWang/UCSCXenaTools/issues?utf8=%E2%9C%93&q=is%3Aissue+is%3Aopen+)
+<thead>
+
+<tr class="header">
+
+<th align="left">
+
+Usage
+
+</th>
+
+<th align="left">
+
+Release
+
+</th>
+
+<th align="left">
+
+Development
+
+</th>
+
+</tr>
+
+</thead>
+
+<tbody>
+
+<tr class="odd">
+
+<td align="left">
+
+<a href="https://www.gnu.org/licenses/gpl-3.0.en.html"><img src="https://img.shields.io/badge/licence-GPL--3-blue.svg" alt="Licence"></a>
+
+</td>
+
+<td align="left">
+
+<a href="https://cran.r-project.org/package=UCSCXenaTools"><img src="https://www.r-pkg.org/badges/version/UCSCXenaTools" alt="CRAN"></a>
+
+</td>
+
+<td align="left">
+
+<a href="https://travis-ci.org/ShixiangWang/UCSCXenaTools"><img src="https://travis-ci.org/ShixiangWang/UCSCXenaTools.svg?branch=master" alt="Travis"></a>
+
+</td>
+
+</tr>
+
+<tr class="even">
+
+<td align="left">
+
+<a href="https://cran.r-project.org/"><img src="https://img.shields.io/badge/R%3E%3D-3.0.0-blue.svg" alt="minimal R version"></a>
+
+</td>
+
+<td align="left">
+
+<a href="https://cran.r-project.org/web/checks/check_results_UCSCXenaTools.html"><img src="https://cranchecks.info/badges/summary/UCSCXenaTools" alt="cran-checks"></a>
+
+</td>
+
+<td align="left">
+
+<a href="https://ci.appveyor.com/project/ShixiangWang/UCSCXenaTools"><img src="https://ci.appveyor.com/api/projects/status/github/ShixiangWang/UCSCXenaTools?branch=master&svg=true" alt="AppVeyor"></a>
+
+</td>
+
+</tr>
+
+<tr class="odd">
+
+<td align="left">
+
+<a href="https://CRAN.R-project.org/package=UCSCXenaTools"><img src="https://tinyverse.netlify.com/badge/UCSCXenaTools"></a>
+
+</td>
+
+<td align="left">
+
[…]
