### 2019-06-19 a15ef7a
add paper template
