### 2019-06-20 13944b6
Update fetch.Rd
