### 2019-06-20 #315 @melvidoni
### Editor checks:

---

#### Editor comments

Thank you for submitting your package! Please add a LICENCE file to the repo, and fix the following suggestions of `goodpractices::gp()` package:

```
── GP UCSCXenaTools ──────────────────────────────────────────────────────────────────────────

It is good practice to

  ✖ not import packages as a whole, as this can cause name clashes
    between the imported packages. Instead, import only the specific functions you
    need.
──────────────────────────────────────────────────────────────────────────────────────────────
```

---

Reviewers: TBD
Due date: TBD
