### 2019-06-21 #315 @ShixiangWang
Dear editor @melvidoni,

All the comments you mentioned above have been fixed. A [GPL-3 license file](https://github.com/ShixiangWang/UCSCXenaTools/blob/master/LICENSE.md) has been added by `usethis::use_gpl3_license` command. All `import` sentences in [NAMESPACE file](https://github.com/ShixiangWang/UCSCXenaTools/blob/master/NAMESPACE) have been modified to `importFrom`, and [corresponding code](https://github.com/ShixiangWang/UCSCXenaTools/commit/0b374bfbfdaee70d32791609ae76293f8f935e5f) have also been modified.

Thanks for your comments. Please let me know if there are other issues.

Best,
Shixiang
