+++ README.Rmd
@@ -96,8 +96,8 @@ Development
 
 # UCSCXenaTools <img src='man/figures/logo.png' align="right" height="200" alt="logo"/>
 
-**UCSCXenaTools** is an R package downloading and exploring data from
-[**UCSC Xena data hubs**](https://xenabrowser.net/datapages/), which are a collection of UCSC-hosted public databases such as TCGA, ICGC, TARGET, GTEx, CCLE, and others. Databases are normalized so they can be combined, linked, filtered, explored and downloaded.
+**UCSCXenaTools** is an R package for accessing genomics data from UCSC Xena platform, from cancer multi-omics to single-cell RNA-seq. 
+Public omics data for UCSC Xena are supported through [**multiple turn-key Xena Hubs**](https://xenabrowser.net/datapages/), which are a collection of UCSC-hosted public databases such as TCGA, ICGC, TARGET, GTEx, CCLE, and others. Databases are normalized so they can be combined, linked, filtered, explored and downloaded.
 
 ## Table of Contents
 
@@ -251,10 +251,19 @@ If you want to use APIs provided by **UCSCXenaTools** to access Xena Hubs, pleas
 
 ## Citation
 
+Cite me by the following paper.
+
 ```
 Wang, Shixiang, et al. "The predictive power of tumor mutational burden 
     in lung cancer immunotherapy response is influenced by patients' sex." 
     International journal of cancer (2019).
+```
+
+Cite UCSC Xena by the following paper. 
+
+```
+Goldman, Mary, et al. "The UCSC Xena Platform for cancer genomics data 
+    visualization and interpretation." BioRxiv (2019): 326470.
 ```
 
 ## Acknowledgments
