+++ README.md
@@ -150,11 +150,13 @@ Thanks\!](https://img.shields.io/badge/say-thanks-ff69b4.svg)](https://saythanks
 
 # UCSCXenaTools <img src='man/figures/logo.png' align="right" height="200" alt="logo"/>
 
-**UCSCXenaTools** is an R package downloading and exploring data from
-[**UCSC Xena data hubs**](https://xenabrowser.net/datapages/), which are
-a collection of UCSC-hosted public databases such as TCGA, ICGC, TARGET,
-GTEx, CCLE, and others. Databases are normalized so they can be
-combined, linked, filtered, explored and downloaded.
+**UCSCXenaTools** is an R package for accessing genomics data from UCSC
+Xena platform, from cancer multi-omics to single-cell RNA-seq. Public
+omics data for UCSC Xena are supported through [**multiple turn-key Xena
+Hubs**](https://xenabrowser.net/datapages/), which are a collection of
+UCSC-hosted public databases such as TCGA, ICGC, TARGET, GTEx, CCLE, and
+others. Databases are normalized so they can be combined, linked,
+filtered, explored and downloaded.
 
 ## Table of Contents
 
@@ -284,7 +286,7 @@ Query and download.
 XenaQuery(df_todo) %>%
   XenaDownload() -> xe_download
 #> This will check url status, please be patient.
-#> All downloaded files will under directory /var/folders/mx/rfkl27z90c96wbmn3_kjk8c80000gn/T//RtmpLiBqVs.
+#> All downloaded files will under directory /var/folders/mx/rfkl27z90c96wbmn3_kjk8c80000gn/T//RtmpAAOXMJ.
 #> The 'trans_slash' option is FALSE, keep same directory structure as Xena.
 #> Creating directories for datasets...
 #> Downloading TCGA.LUAD.sampleMap/LUAD_clinicalMatrix.gz
@@ -389,9 +391,18 @@ vignette](https://shixiangwang.github.io/home/en/tools/ucscxenatools-api/).
 
 ## Citation
 
+Cite me by the following
+    paper.
+
     Wang, Shixiang, et al. "The predictive power of tumor mutational burden 
         in lung cancer immunotherapy response is influenced by patients' sex." 
         International journal of cancer (2019).
+
+Cite UCSC Xena by the following
+    paper.
+
+    Goldman, Mary, et al. "The UCSC Xena Platform for cancer genomics data 
+        visualization and interpretation." BioRxiv (2019): 326470.
 
 ## Acknowledgments
 
