### 2019-06-21 4a2811b
prepare paper header
