+++ README.Rmd
@@ -176,6 +176,21 @@ XenaGenerate(subset = XenaHostNames=="tcgaHub") %>%
 df_todo
 ```
 
+Sometimes we only know some keywords, `XenaScan()` can be used to scan all rows to detect if 
+the keywords exist in `XenaData`.
+
+
+```{r}
+x1 = XenaScan(pattern = 'Blood')
+x2 = XenaScan(pattern = 'LUNG', ignore.case = FALSE)
+
+x1 %>%
+    XenaGenerate()
+x2 %>%
+    XenaGenerate()
+```
+
+
 Query and download.
 
 ```{r}
