+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: UCSCXenaTools
 Title: Download and Explore Datasets from UCSC Xena Data Hubs
-Version: 1.2.3
+Version: 1.2.4
 Authors@R: c(person("Shixiang", "Wang", email = "w_shixiang@163.com", role = c("aut", "cre"), 
             comment = c(ORCID = "0000-0001-9855-7357")),
             person("Martin", "Morgan", role="aut"))
@@ -19,6 +19,7 @@ Imports:
     dplyr,
     magrittr,
     jsonlite,
+    rlang,
     shinydashboard,
     shiny,
     DT
