+++ NAMESPACE
@@ -10,6 +10,7 @@ export(XenaHub)
 export(XenaPrepare)
 export(XenaQuery)
 export(XenaQueryProbeMap)
+export(XenaScan)
 export(XenaShiny)
 export(availTCGA)
 export(cohorts)
