+++ tests/testthat/test-basic-workflow.R
@@ -19,6 +19,7 @@ test_that("XenaGenerate works", {
 
 test_that("Xena Scan works", {
   XenaScan()
+  XenaScan(pattern = "zzzzzzzzzzz")
   x1 = XenaScan(pattern = 'Blood')
   x2 = XenaScan(pattern = 'LUNG', ignore.case = FALSE)
 
