+++ NAMESPACE
@@ -42,6 +42,7 @@ importFrom(methods,slot)
 importFrom(readr,cols)
 importFrom(readr,read_tsv)
 importFrom(readr,read_tsv_chunked)
+importFrom(rlang,.data)
 importFrom(shiny,column)
 importFrom(shiny,fluidRow)
 importFrom(shiny,icon)
