### 2019-07-22 #315 @ChristineStawitz-NOAA
Apologies for getting this in at the last possible moment!

    **This might be obvious to those who use the USCS data hubs, but the target audience was not clear to me in the README; please add in this information as described above in the README.**

  **Worked fine on my PC**

1. I was able to find the Vignettes on the author's website. However, they are not installed locally either when using `devtools::install_github()` or installing from CRAN. Even if there is a website version that is more up-to-date, it seems like a local version would be useful, particularly for CRAN users who might not find the Github link.
2. I tried some of the intro vignette with the CRAN version and some with the Github one - with the CRAN version, `XenaScan()` is not found. There should be some guidance on minimum versions to run the vignette when describing installation.
3. Intro vignette - For the examples using local files with `XenaPrepare()` it would be better to put some sample .gz files in the `data/` directory and use  rather than referencing `E:/Github` or `./downloads/` which all users won't have. I'm not sure if this is recommended by rOpenSci but in the past I've used `system.file("data", package = "UCSCXenaTools")` to access this directory.
4. API vignette - typo in one of the headers: "Find out the number of idnetifiers in a dataset"

  **There is no guidance on how to contribute**

>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

 **I get a warning when running `devtools::check()` - "qpdf is needed for size reduction on PDFs"**

1. Argument naming is not consistent - for example, the `fetch_` API functions are in snake_case but a number of other functions are camel-cased with the first letter capitalized (i.e. `XenaScan`) or regular camel-cased (`getTCGAdata`). It would be helpful to adopt a similar casing style even if snake_case doesn't work because of consistency with other tools for these data sets.
2. The Workflow.R and simplify.R files are a bit overloaded - style guidelines suggest not putting all of the key functions in a single file, and it wasn't clear to me where to search for some of the different functions.
3. Throughout the R code, `::` is used to reference functions within the `UCSCXenaTools` package itself as well as `dplyr`, which is imported. It could clean things up to remove these for packages that are explicitly imported.
4. I think `shiny` and `shinydashboard` could be moved to suggests instead of imports in the `DESCRIPTION`

---

I'm not in the omics field, but this seems like a useful package for interacting with the UCSC Xena platform. I found the package reliable and stable, but it was a bit hard to follow where the vignettes were located after local install and there were a few vignette issues (see above). There are a couple of code simplifications and consistency improvements that could be made as well, which are detailed above.
