+++ README.Rmd
@@ -99,15 +99,18 @@ Development
 **UCSCXenaTools** is an R package for accessing genomics data from UCSC Xena platform, from cancer multi-omics to single-cell RNA-seq. 
 Public omics data from UCSC Xena are supported through [**multiple turn-key Xena Hubs**](https://xenabrowser.net/datapages/), which are a collection of UCSC-hosted public databases such as TCGA, ICGC, TARGET, GTEx, CCLE, and others. Databases are normalized so they can be combined, linked, filtered, explored and downloaded.
 
+**Who is the target audience and what are scientific applications of this package?**
+
+* Target Audience: cancer and clinical researchers, bioinformaticians
+* Applications: genomic and clinical analyses
+
 ## Table of Contents
 
 * [Installation](#installation)
 * [Data Hub List](#data-hub-list)
-* [Usage](#usage)
-* [Documentation](#documentation)
-* [APIs](#apis)
+* [Basic usage](#basic-usage)
 * [Citation](#citation)
-* [Acknowledgement](#acknowledgement)
+* [Acknowledgment](#acknowledgment)
 
 ## Installation
 
@@ -144,11 +147,11 @@ Currently, **UCSCXenaTools** supports 10 data hubs of UCSC Xena.
 If any url of data hub is changed or a new data hub is online, please remind me by emailing to <w_shixiang@163.com> or [opening an issue on GitHub](https://github.com/ShixiangWang/UCSCXenaTools/issues).
 
 
-## Usage
+## Basic usage
 
 Download UCSC Xena datasets and load them into R by **UCSCXenaTools** is a workflow with `generate`, `filter`, `query`, `download` and `prepare` 5 steps, which are implemented as `XenaGenerate`, `XenaFilter`, `XenaQuery`, `XenaDownload` and `XenaPrepare` functions, respectively. They are very clear and easy to use and combine with other packages like `dplyr`.
 
-To show the basic usage of **UCSCXenaTools**, we will download clinical data of LUNG, LUAD, LUSC from TCGA (hg19 version) data hub.
+To show the basic usage of **UCSCXenaTools**, we will download clinical data of LUNG, LUAD, LUSC from TCGA (hg19 version) data hub. Users can learn more about **UCSCXenaTools** by running `browseVignettes("UCSCXenaTools")` to read vignette.
 
 ### XenaData data.frame
 
@@ -176,21 +179,6 @@ XenaGenerate(subset = XenaHostNames=="tcgaHub") %>%
 df_todo
 ```
 
-Sometimes we only know some keywords, `XenaScan()` can be used to scan all rows to detect if 
-the keywords exist in `XenaData`.
-
-
-```{r}
-x1 = XenaScan(pattern = 'Blood')
-x2 = XenaScan(pattern = 'LUNG', ignore.case = FALSE)
-
-x1 %>%
-    XenaGenerate()
-x2 %>%
-    XenaGenerate()
-```
-
-
 Query and download.
 
 ```{r}
@@ -205,64 +193,6 @@ cli = XenaPrepare(xe_download)
 class(cli)
 names(cli)
 ```
-
-
-### Browse datasets
-
-Create two XenaHub objects:
-
-* `to_browse` - a XenaHub object containing a cohort and a dataset.
-* `to_browse2` - a XenaHub object containing 2 cohorts and 2 datasets.
-
-```{r}
-XenaGenerate(subset = XenaHostNames=="tcgaHub") %>%
-    XenaFilter(filterDatasets = "clinical") %>%
-    XenaFilter(filterDatasets = "LUAD") -> to_browse
-
-to_browse
-
-XenaGenerate(subset = XenaHostNames=="tcgaHub") %>%
-    XenaFilter(filterDatasets = "clinical") %>%
[…]
