+++ NEWS.md
@@ -1,3 +1,9 @@
+# v1.2.5
+
+## Changes
+
+* Update README, vignette and code according to review comments from Christine Stawitz, detail at <https://github.com/ropensci/software-review/issues/315>
+
 # v1.2.4
 
 ## New features
