+++ NAMESPACE
@@ -43,21 +43,6 @@ importFrom(readr,cols)
 importFrom(readr,read_tsv)
 importFrom(readr,read_tsv_chunked)
 importFrom(rlang,.data)
-importFrom(shiny,column)
-importFrom(shiny,fluidRow)
-importFrom(shiny,icon)
-importFrom(shiny,runApp)
-importFrom(shiny,selectInput)
-importFrom(shiny,shinyApp)
-importFrom(shiny,updateSelectInput)
-importFrom(shinydashboard,dashboardBody)
-importFrom(shinydashboard,dashboardHeader)
-importFrom(shinydashboard,dashboardPage)
-importFrom(shinydashboard,dashboardSidebar)
-importFrom(shinydashboard,menuItem)
-importFrom(shinydashboard,sidebarMenu)
-importFrom(shinydashboard,tabItem)
-importFrom(shinydashboard,tabItems)
 importFrom(utils,URLencode)
 importFrom(utils,browseURL)
 importFrom(utils,download.file)
