+++ R/workflow.R
@@ -1,647 +1,0 @@
-# FUN: XenaHub (download) workflow
-
-##' Generate and Subset a XenaHub Object from 'XenaData'
-##'
-##' @author Shixiang Wang <w_shixiang@163.com>
-##' @param XenaData a `data.frame`. Default is `data(XenaData)`.
-##' The input of this option can only be `data(XenaData)` or its subset.
-##' @param subset logical expression indicating elements or rows to keep.
-##' @return a [XenaHub] object.
-##' @export
-##' @examples
-##' # 1 get all datasets
-##' XenaGenerate()
-##' # 2 get TCGA BRCA
-##' XenaGenerate(subset = XenaCohorts == "TCGA Breast Cancer (BRCA)")
-##' # 3 get all datasets containing BRCA
-##' XenaGenerate(subset = grepl("BRCA", XenaCohorts))
-
-XenaGenerate <- function(XenaData = UCSCXenaTools::XenaData,
-                         subset = TRUE) {
-  enclos <- parent.frame()
-  subset <- substitute(subset)
-  row_selector <- eval(subset, XenaData, enclos)
-  XenaData <- XenaData[row_selector, ]
-
-  .XenaHub(
-    hosts = unique(XenaData$XenaHosts),
-    cohorts = unique(XenaData$XenaCohorts),
-    datasets = XenaData$XenaDatasets
-  )
-}
-
-##' Filter a XenaHub Object
-##'
-##' One of main functions in **UCSCXenatools**. It is used to filter
-##' `XenaHub` object according to cohorts, datasets. All datasets can be found
-##' at <https://xenabrowser.net/datapages/>.
-##'
-##' @param x a [XenaHub] object
-##' @param filterCohorts default is `NULL`. A character used to filter cohorts,
-##' regular expression is supported.
-##' @param filterDatasets default is `NULL`. A character used to filter datasets,
-##' regular expression is supported.
-##' @param ignore.case if `FALSE`, the pattern matching is case sensitive
-##' and if `TRUE`, case is ignored during matching.
-##' @param ... other arguments except `value` passed to [base::grep()].
-##' @return a `XenaHub` object
-##' @author Shixiang Wang <w_shixiang@163.com>
-##' @export
-##' @examples
-##' # operate TCGA datasets
-##' xe = XenaGenerate(subset = XenaHostNames == "tcgaHub")
-##' xe
-##' # get all names of clinical data
-##' xe2 = XenaFilter(xe, filterDatasets = "clinical")
-##' datasets(xe2)
-XenaFilter <- function(x,
-                       filterCohorts = NULL,
-                       filterDatasets = NULL,
-                       ignore.case = TRUE, ...) {
-  if (is.null(filterCohorts) & is.null(filterDatasets)) {
-    warning("No operation for input, do nothing...")
-    return(x)
-  }
-
-  cohorts_select <- character()
-  datasets_select <- character()
-
-  # suppress binding notes
-  XenaHosts <- XenaCohorts <- XenaDatasets <- NULL
-
-  if (!is.null(filterCohorts)) {
-    cohorts_select <- grep(
-      pattern = filterCohorts,
-      x@cohorts,
-      ignore.case = ignore.case,
-      value = TRUE,
-      ...
-    )
-  }
-
-  if (!is.null(filterDatasets)) {
-    datasets_select <- grep(
-      pattern = filterDatasets,
-      x@datasets,
-      ignore.case = ignore.case,
-      value = TRUE,
-      ...
-    )
-  }
-
-  if (identical(cohorts_select, character()) &
-    identical(datasets_select, character())) {
-    warning("No valid cohorts or datasets find! Please check your input.")
-  } else {
[…]
