+++ man/XenaQuery.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/workflow.R
+% Please edit documentation in R/XenaQuery.R
 \name{XenaQuery}
 \alias{XenaQuery}
 \title{Query URL of Datasets before Downloading}

[…]
