### 2019-07-23 #315 @ShixiangWang
@ChristineStawitz-NOAA  Thanks for your all comments. Basically, I fixed all the problems you raised. And I also have some reasons for the non-modified parts.

All replies for problematic points have been put as the following. Please forgive me my English if you find something hard to understand, just ask me again. :)

  Thanks, this information has been added to [README](https://github.com/ShixiangWang/UCSCXenaTools/)

* [x] **Vignette(s)**

  1. I have created a [local version](https://github.com/ShixiangWang/UCSCXenaTools/blob/master/vignettes/USCSXenaTools.Rmd) describing the basic usage of **UCSCXenaTools**, you can install it from the latest version at GitHub by `remotes::install_github("ShixiangWang/UCSCXenaTools", build_vignettes = TRUE)`
  2. The minimum version has been described at [installation section](https://github.com/ShixiangWang/UCSCXenaTools/blob/88f4e9ff0971ed798cfddec6df8bb964396d841a/vignettes/USCSXenaTools.Rmd#L62)
  3. I have updated the intro vignette and use /tmp directory to describe the message. I don't use local data because it will increase package size dramatically.
  4. The typo has been fixed, thanks :).

* [ ] **Submitting to JOSS**

	I don't know what should I do next to submit the package to JOSS. After reviewing, is this package accepted by JOSS automatically? Or should I write the summary, authors, etc.. now and submit it along with the package to JOSS?

* [x] **Test**

I get a warning when running devtools::check() - "qpdf is needed for size reduction on PDFs"

* `qpdf` is required for R package check but not this package, after you install `qpdf` according to your system, the warning will be gone.

* [x] **Packaging**
	1. Argument naming is indeed not consistent.  The naming strategy has a different purpose for now. Functions start with `Xena` are core functions which can be combined by pipe operator `%>%` to operate datasets;  functions like `getTCGAdata`, `avail*` etc. are side effects that I don't want the users to use them but they are useful for some users; functions using snake case basically are API functions for interacting with UCSC Xena database to query/download a part of a dataset. I agree with your consistent idea, but for now, I can't adjust them before I prepare to make an incompatible version.
	2. Thanks and I have split the `workflow.R` into [individual files](https://github.com/ShixiangWang/UCSCXenaTools/tree/master/R). The functions in`simplify.R` all work for one purpose: select and download TCGA datasets, it will be harder to maintain if I split it up.
	3. The mentioned code I don't like to change because they make me clearly know where the data or functions come from and write less redundant document comments like `@importFrom`.
	4. I have moved `shiny`, `shinydashborad` and `DT` packages to suggests and cleaned [the corresponding code](https://github.com/ShixiangWang/UCSCXenaTools/blob/master/R/shiny.R). This reduces the core dependencies, thanks.
