### 2019-07-23 847d8ea
Update NEWS.md
