### 2019-07-23 ropensci/UCSCXenaTools#11 @ShixiangWang
Function naming strategy
Comments from https://github.com/ropensci/software-review/issues/315

* Argument naming is not consistent - for example, the fetch_ API functions are in snake_case but a number of other functions are camel-cased with the first letter capitalized (i.e. XenaScan) or regular camel-cased (getTCGAdata). It would be helpful to adopt a similar casing style even if snake_case doesn't work because of consistency with other tools for these data sets.

This issue should be fixed in the next incompatible version.
