### 2019-07-24 #315 @carlganz
>##### For packages co-submitting to JOSS
>
>- [X] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [X] **A short summary** describing the high-level functionality of the software
>- [X] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [X] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

I have very little knowledge of genomics, but I can see clearly that the UCSC Xena platform is valuable to the genomics community, and after thoroughly reviewing the code within your package I can see that working with the Xena platform is highly non-trivial, and that this package encapsulates a great deal of work that succesfully makes the Xena platform much more accessible to R users. Congrats on the wonderful package!

# Statement of Need and JOSS summary

You explain what the Xena platform is, but it would be nice to have a sentence or two explaining why interfacing with the Xena platform is non-trivial, and how this interface significantly improves R user access.

* When I used `remotes::install_github("ShixiangWang/UCSCXenaTools", build_vignettes = TRUE)` I got error about missing `prettydoc` dependency, but I found `remotes::install_github("ShixiangWang/UCSCXenaTools", build_vignettes = TRUE, dependencies = TRUE)` worked fine.

# Small code suggestions

* [Here](https://github.com/ShixiangWang/UCSCXenaTools/blob/master/R/XenaHub-class.R#L215) and a few other places use the discouraged antipattern `1:length(some_vec)` or `1:nrow(some_df)`. `seq_len` is the safer alternative, but I am pretty sure it won't make a difference in this case.

* Another antipattern worth avoiding is iteratively building a data.frame with `rbind` in a loop like [here](https://github.com/ShixiangWang/UCSCXenaTools/blob/master/R/XenaHub-class.R#L244). Section 4 in [this blog post by @noamross](https://www.noamross.net/blog/2014/4/16/vectorization-in-r--why.html) explains why not to do this, but essentially each iteration copys the entire data.frame in memory.

So I would replace this:

```r
  resDF <- data.frame(stringsAsFactors = FALSE)
  for (i in 1:length(XenaList)) {
    hostNames <- names(XenaList)[i]
    cohortNames <- names(XenaList[[i]])
    res <- data.frame(stringsAsFactors = FALSE)

    for (j in 1:length(cohortNames)) {
      oneCohort <- XenaList[[i]][j]
      # The unassigned cohorts have NULL data, remove it
      if (names(oneCohort) != "(unassigned)") {
        resCohort <- data.frame(
          XenaCohorts = names(oneCohort),
          XenaDatasets = as.character(oneCohort[[1]]),
          stringsAsFactors = FALSE
        )
        res <- rbind(res, resCohort)
      }
    }
    res$XenaHosts <- hostNames
    resDF <- rbind(resDF, res)
  }
```
With something like:
```r
  # use lapply to create list of dfs because why introduce purrr dependency
  resDF <- lapply(seq_len(XenaList), function(i) {
    hostNames <- names(XenaList)[i]
    cohortNames <- names(XenaList[[i]])
    res <- data.frame(stringsAsFactors = FALSE)

    for (j in 1:length(cohortNames)) {
      oneCohort <- XenaList[[i]][j]
      # The unassigned cohorts have NULL data, remove it
      if (names(oneCohort) != "(unassigned)") {
        resCohort <- data.frame(
          XenaCohorts = names(oneCohort),
          XenaDatasets = as.character(oneCohort[[1]]),
          stringsAsFactors = FALSE
        )
        res <- rbind(res, resCohort)
      }
    }
    res$XenaHosts <- hostNames
    return(res)
  }) %>%
  # convert list of dfs into single df
  bind_rows
```

* Documentation from dplyr says `dplyr::rbind_list` is deprecated and to use `dplyr::bind_rows`, which I recommend you do [here](https://github.com/ShixiangWang/UCSCXenaTools/blob/master/R/XenaHub-class.R#L308) just in case Hadley removes the function at some point you don't want to be a breaking reverse dependency on CRAN.

Hope that helps!
