+++ paper/paper.bib
@@ -6,3 +6,64 @@
   year={2019},
   publisher={Cold Spring Harbor Laboratory}
 }
+
+@article{weinstein2013cancer,
+  title={The cancer genome atlas pan-cancer analysis project},
+  author={Weinstein, John N and Collisson, Eric A and Mills, Gordon B and Shaw, Kenna R Mills and Ozenberger, Brad A and Ellrott, Kyle and Shmulevich, Ilya and Sander, Chris and Stuart, Joshua M and Cancer Genome Atlas Research Network and others},
+  journal={Nature genetics},
+  volume={45},
+  number={10},
+  pages={1113},
+  year={2013},
+  publisher={Nature Publishing Group}
+}
+
+@article{zhang2011international,
+  title={International Cancer Genome Consortium Data Portal—a one-stop shop for cancer genomics data},
+  author={Zhang, Junjun and Baran, Joachim and Cros, Anthony and Guberman, Jonathan M and Haider, Syed and Hsu, Jack and Liang, Yong and Rivkin, Elena and Wang, Jianxin and Whitty, Brett and others},
+  journal={Database},
+  volume={2011},
+  year={2011},
+  publisher={Oxford University Press}
+}
+
+@article{barretina2012cancer,
+  title={The Cancer Cell Line Encyclopedia enables predictive modelling of anticancer drug sensitivity},
+  author={Barretina, Jordi and Caponigro, Giordano and Stransky, Nicolas and Venkatesan, Kavitha and Margolin, Adam A and Kim, Sungjoon and Wilson, Christopher J and Leh{\'a}r, Joseph and Kryukov, Gregory V and Sonkin, Dmitriy and others},
+  journal={Nature},
+  volume={483},
+  number={7391},
+  pages={603},
+  year={2012},
+  publisher={Nature Publishing Group}
+}
+
+@article{mullighan2008genomic,
+  title={Genomic analysis of the clonal origins of relapsed acute lymphoblastic leukemia},
+  author={Mullighan, Charles G and Phillips, Letha A and Su, Xiaoping and Ma, Jing and Miller, Christopher B and Shurtleff, Sheila A and Downing, James R},
+  journal={Science},
+  volume={322},
+  number={5906},
+  pages={1377--1380},
+  year={2008},
+  publisher={American Association for the Advancement of Science}
+}
+
+@article{puram2017single,
+  title={Single-cell transcriptomic analysis of primary and metastatic tumor ecosystems in head and neck cancer},
+  author={Puram, Sidharth V and Tirosh, Itay and Parikh, Anuraag S and Patel, Anoop P and Yizhak, Keren and Gillespie, Shawn and Rodman, Christopher and Luo, Christina L and Mroz, Edmund A and Emerick, Kevin S and others},
+  journal={Cell},
+  volume={171},
+  number={7},
+  pages={1611--1624},
+  year={2017},
+  publisher={Elsevier}
+}
+
+@article{wickham2015dplyr,
+  title={dplyr: A grammar of data manipulation},
+  author={Wickham, Hadley and Francois, Romain and Henry, Lionel and M{\"u}ller, Kirill and others},
+  journal={R package version 0.4},
+  volume={3},
+  year={2015}
+}
