+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'UCSCXenaTools: an R package for accessing genomics data from UCSC Xena platform'
+title: 'UCSCXenaTools: an R package for Accessing Genomics Data from UCSC Xena platform, from Cancer Multi-omics to Single-cell RNA-seq'
 tags:
   - R
   - cancer genomics
@@ -18,33 +18,51 @@ affiliations:
    index: 2
  - name: University of Chinese Academy of Sciences
    index: 3
-date: 21 June 2019
+date: 24 July 2019
 bibliography: paper.bib
 ---
 
 # Summary
 
-UCSC Xena is a visual exploration resource for both public and private omics data, supported through the web-based Xena Browser and multiple turn-key Xena Hubs. This unique archecture allows researchers to view their own data securely, using private Xena Hubs, simultaneously visualizing large public cancer genomics datasets, including TCGA and the GDC. Data integration occurs only within the Xena Browser, keeping private data private. Xena supports virtually any functional genomics data, including SNVs, INDELs, large structural variants, CNV, expression, DNA methylation, ATAC-seq signals, and phenotypic annotations. Browser features include the Visual Spreadsheet, survival analyses, powerful filtering and subgrouping, statistical analyses, genomic signatures, and bookmarks. Xena differentiates itself from other genomics tools, including its predecessor, the UCSC Cancer Genomics Browser, by its ability to easily and securely view public and private data, its high performance, its broad data type support, and many unique features.
+UCSC Xena platform (https://xenabrowser.net/) provides unprecedented resource for public omics data [@goldman2019ucsc]
+from big projects like The Cancer Genome Atlas (TCGA) [@weinstein2013cancer], 
+International Cancer Genome Consortium Data Portal (ICGC) [@zhang2011international],
+The Cancer Cell Line Encyclopedia (CCLE) [@barretina2012cancer], or reserach groups like @mullighan2008genomic, @puram2017single.
+All available data types include single-nucleotide variants (SNVs), small insertions and deletions (INDELs), large structural variants, copy number variation (CNV), expression, DNA methylation, ATAC-seq signals, and phenotypic annotations. 
+
+Despite UCSC Xena platform itself allows users to explore and analyze data, it is hard
+for users to incorporate multiple datasets or data types, integrate the selected data with 
+popular analysis tools or homebrewed code, and reproduce analysis procedures.
+R language is well established and extensively used standard in statistical and bioinformatics research.
+Here, we introduce an R package UCSCXenaTools for enabling data retrieval, analysis integration and 
+reproducible research for omics data from UCSC Xena platform.
+
+Currently, UCSCXenaTools supports downloading over 1600 datasets from 10 data hubs of UCSC Xena platform
+as shown in Table 1. Downloading UCSC Xena datasets and loading them into R by UCSCXenaTools is a workflow with generate, filter, query, download and prepare 5 steps, which are implemented as functions XenaGenerate, XenaFilter, XenaQuery, XenaDownload and XenaPrepare, respectively. They are very clear and easy to use and combine with other packages like dplyr [@wickham2015dplyr].
+Besides, UCSCXenaTools can also query and download subset of a target dataset, 
+this is particularly useful when
+user focus on studying an object like gene or protein. The key features are summarized in Figure 1.
 
 
-# Citations
+|Data hub       | Dataset count|URL                                |
+|:--------------|-------------:|:----------------------------------|
+|tcgaHub        |           879|https://tcga.xenahubs.net          |
[…]
