+++ vignettes/USCSXenaTools.Rmd
@@ -47,7 +47,13 @@ You can also install devel version of **UCSCXenaTools** from github with:
 
 ```{r gh-installation, eval = FALSE}
 # install.packages("remotes")
-remotes::install_github("ShixiangWang/UCSCXenaTools", build_vignettes = TRUE)
+remotes::install_github("ShixiangWang/UCSCXenaTools")
+```
+
+If you want to build vignette in local, please add two options:
+
+```{r, eval=FALSE}
+remotes::install_github("ShixiangWang/UCSCXenaTools", build_vignettes = TRUE, dependencies = TRUE)
 ```
 
 The minimum versions to run the vignette is `1.2.4`. 

[…]
