### 2019-07-24 #315 @ShixiangWang
@ChristineStawitz-NOAA Sorry for ignoring  `There is no guidance on how to contribute` yesterday, I have added it [here](https://github.com/ShixiangWang/UCSCXenaTools#how-to-contribute).

@carlganz Thanks for your comments :). The modification can be seen from the last [two commits](https://github.com/ShixiangWang/UCSCXenaTools/commits/master).  I also reply your comments point to point as the following.

#### Statement of Need and JOSS summary

It was a template in repository yesterday. I rewrite it today, you can see new version [here](https://github.com/ShixiangWang/UCSCXenaTools/blob/master/paper/paper.md). I am not good at English and also writing, so any suggestion is welcome :).

#### Installation

**prettydoc** package is in suggests field, so it will not be install automatically without option `dependencies=TRUE`. Thank   you for poiniting it out, I have added this message in [new README](https://github.com/ShixiangWang/UCSCXenaTools)

#### use seq_len

I like `1:length(x)` this format because it is more easy to read. `seq_len` is indeed a safer version for object with length 0. I have changed  corresponding code.

#### use for

I won’t change such code for now. I do like `for` more than `lapply` here because `for` loop is much easier to read and debug.

For example,

```
  # use lapply to create list of dfs because why introduce purrr dependency
  resDF <- lapply(seq_len(XenaList), function(i) {
    hostNames <- names(XenaList)[i]
    cohortNames <- names(XenaList[[i]])
    res <- data.frame(stringsAsFactors = FALSE)
    ...
```

The code won’t work before modifying all `XenaList` to `i` and removing `[i]`. And I have to debug to make sure the code works.

When I developed this package, I have no idea how to use `purrr`, so this package is not used here.

As Hadley wrote in *Advance in R*, the for loop in R now is very quick. It will not affect the performance.

#### use bind_rows

This is modified.
