### 2019-07-24 #315 @ShixiangWang
I am glad to add two reviewers in DESCRIPTION file according to

 > Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer ("rev" role) in the package DESCRIPTION file.

Thanks all your comments @ChristineStawitz-NOAA @carlganz
