### 2019-07-25 #315 @ChristineStawitz-NOAA
**Statement of Need and JOSS paper**
I don't see the bibliography showing up at the end of the paper.md file, though I can see the in-text references.

I initially did not review this as a JOSS submission, should I be also following their review criteria @melvidoni or will it go through a separate review at JOSS after the rOpenSci one?

I have some long-running code going in other `RStudio` sessions so I will not be able to reinstall the package and check the vignette again until tomorrow - assuming that works I will sign off on the review then as I am satisfied with the changes from what is in this thread and the repo :) Nice work @ShixiangWang ! I am also OK with being credited for the review, thanks.
