+++ DESCRIPTION
@@ -3,7 +3,10 @@ Title: Download and Explore Datasets from UCSC Xena Data Hubs
 Version: 1.2.4.9000
 Authors@R: c(person("Shixiang", "Wang", email = "w_shixiang@163.com", role = c("aut", "cre"), 
             comment = c(ORCID = "0000-0001-9855-7357")),
-            person("Martin", "Morgan", role="aut"))
+            person("Martin", "Morgan", role="ctb"),
+            person("Xue-Song", "Liu", role = "ctb", comment = c(ORCID = "0000-0002-7736-0077")),
+            person("Christine", "Stawitz", role="rev"),
+            person("Carl", "Ganz", role="rev"))
 Maintainer: Shixiang Wang <w_shixiang@163.com>
 URL: https://github.com/ShixiangWang/UCSCXenaTools
 BugReports: https://github.com/ShixiangWang/UCSCXenaTools/issues
