+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ## Changes
 
+* updated author list
 * updated README, vignette and code according to review comments from Christine Stawitz, detail at <https://github.com/ropensci/software-review/issues/315>
 * moved some packages to suggests
 * splitted `workflow.R` into multiple files
