### 2019-07-26 #315 @ShixiangWang
@ChristineStawitz-NOAA the bibliography is shown in [paper.bib](https://github.com/ShixiangWang/UCSCXenaTools/blob/master/paper/paper.bib) file according to [JOSS instruction](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography).

@ChristineStawitz-NOAA @carlganz I have added your names [here](https://github.com/ShixiangWang/UCSCXenaTools/blob/66eba83777e61b5f1a6f2b375b2d2c598cfb8bbf/DESCRIPTION#L8) and [here](https://github.com/ShixiangWang/UCSCXenaTools/blob/master/paper/paper.md#acknowledgements), please take a look and make sure they are right.
