### 2019-07-26 #315 @ChristineStawitz-NOAA
Looks good and I am able to successfully run everything in the vignette
after the changes - thanks!

On Thu, Jul 25, 2019 at 7:27 PM Shixiang Wang <notifications@github.com>
wrote:

> @ChristineStawitz-NOAA <https://github.com/ChristineStawitz-NOAA> the
> bibliography is shown in paper.bib
> <https://github.com/ShixiangWang/UCSCXenaTools/blob/master/paper/paper.bib>
> file according to JOSS instruction
> <https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography>
> .
>
> @ChristineStawitz-NOAA <https://github.com/ChristineStawitz-NOAA>
> @carlganz <https://github.com/carlganz> I have added your names here
> <https://github.com/ShixiangWang/UCSCXenaTools/blob/66eba83777e61b5f1a6f2b375b2d2c598cfb8bbf/DESCRIPTION#L8>
> and here
> <https://github.com/ShixiangWang/UCSCXenaTools/blob/master/paper/paper.md#acknowledgements>,
> please take a look and make sure they are right.
>
