### 2019-08-01 0956934
Modify URLs and tech badges
