+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ## Changes
 
+* moved project to ropensci <https://github.com/ropensci/UCSCXenaTools>
 * updated author list
 * updated README, vignette and code according to review comments from Christine Stawitz, detail at <https://github.com/ropensci/software-review/issues/315>
 * moved some packages to suggests
