+++ DESCRIPTION
@@ -8,8 +8,8 @@ Authors@R: c(person("Shixiang", "Wang", email = "w_shixiang@163.com", role = c("
             person("Christine", "Stawitz", role="rev"),
             person("Carl", "Ganz", role="rev"))
 Maintainer: Shixiang Wang <w_shixiang@163.com>
-URL: https://github.com/ShixiangWang/UCSCXenaTools
-BugReports: https://github.com/ShixiangWang/UCSCXenaTools/issues
+URL: https://github.com/ropensci/UCSCXenaTools
+BugReports: https://github.com/ropensci/UCSCXenaTools/issues
 Description: Download and explore datasets from UCSC Xena data hubs, which are
     a collection of UCSC-hosted public databases such as TCGA, ICGC, TARGET, GTEx, CCLE, and others.
     Databases are normalized so they can be combined, linked, filtered, explored and downloaded.
