### 2019-08-01 #315 @ShixiangWang
@melvidoni Thanks, it works.
There is a problem when I do this step `Activate Zenodo watching the repo`, I can't find my package in the Github list

![image](https://user-images.githubusercontent.com/25057508/62270085-abe42e80-b467-11e9-934d-aa7612881999.png)

It seems I should have rights to access the repositories of ropensci.

![image](https://user-images.githubusercontent.com/25057508/62270197-febde600-b467-11e9-91f3-f0aa98240b92.png)

I don't know where the problem comes from, the Zenodo (I just opened [an issue](https://github.com/zenodo/zenodo/issues/1814)) or the setting from ropensci? Please help me .

I also found I am in a team `auk`, please remove me from it.

![image](https://user-images.githubusercontent.com/25057508/62270311-4b092600-b468-11e9-8a0f-54d1ac52ee6a.png)
