### 2019-08-01 #315 @melvidoni
I removed you from that team @ShixiangWang. Let me try some things regarding Zenodo!
