### 2019-08-02 #315 @melvidoni
Awesome! Then, if you can tackle everything and then let me know, that would be great.
