+++ DESCRIPTION
@@ -3,8 +3,8 @@ Title: Download and Explore Datasets from UCSC Xena Data Hubs
 Version: 1.2.5
 Authors@R: c(person("Shixiang", "Wang", email = "w_shixiang@163.com", role = c("aut", "cre"), 
             comment = c(ORCID = "0000-0001-9855-7357")),
-            person("Martin", "Morgan", role="ctb"),
             person("Xue-Song", "Liu", role = "aut", comment = c(ORCID = "0000-0002-7736-0077")),
+            person("Martin", "Morgan", role="ctb"),
             person("Christine", "Stawitz", role="rev"),
             person("Carl", "Ganz", role="rev"))
 Maintainer: Shixiang Wang <w_shixiang@163.com>
