+++ DESCRIPTION
@@ -5,8 +5,8 @@ Authors@R: c(person("Shixiang", "Wang", email = "w_shixiang@163.com", role = c("
             comment = c(ORCID = "0000-0001-9855-7357")),
             person("Xue-Song", "Liu", role = "aut", comment = c(ORCID = "0000-0002-7736-0077")),
             person("Martin", "Morgan", role="ctb"),
-            person("Christine", "Stawitz", role="rev"),
-            person("Carl", "Ganz", role="rev"))
+            person("Christine", "Stawitz", role="rev", comment = "Christine reviewed the package for ropensci, see <https://github.com/ropensci/software-review/issues/315>"),
+            person("Carl", "Ganz", role="rev", comment = "Carl reviewed the package for ropensci, see <https://github.com/ropensci/software-review/issues/315>"))
 Maintainer: Shixiang Wang <w_shixiang@163.com>
 URL: https://github.com/ropensci/UCSCXenaTools
 BugReports: https://github.com/ropensci/UCSCXenaTools/issues
