### 2019-08-02 #315 @ShixiangWang
@melvidoni I have done everything in your todo list.

I have submitted URL https://doi.org/10.5281/zenodo.3358530 to JOSS.

If the DOI not found, please see https://zenodo.org/record/3358530

@stefaniebutland I would like to write a short introduction to this package.

Thanks all.
Shixiang
