### 2019-08-02 #315 @stefaniebutland
Good to hear that @ShixiangWang

@melvidoni is there an element of the package, it's development, or applications that you think could be highlighted?

For a short introduction, a Tech Note is most appropriate. [Examples](https://ropensci.org/technotes/). It's important not to repeat information that's in the vignette; refer to it where needed. You might illustrate a compelling use case for the package.

Editorial and technical information about preparing your submission, including a template: https://github.com/ropensci/roweb2#contributing-a-blog-post. In the template YAML, for a Tech Note, set `categories: technotes`

Please suggest a deadline for submitting your draft. Don't hesitate to ask me questions. I'll be away Mon Aug 5.

Thank you for doing this extra work. I hope the post will get more eyes on your work.
