### 2019-08-03 #315 @ShixiangWang
@stefaniebutland Thanks, can I submit the draft within a month?

@melvidoni It seems that I should submit repository URL instead of Zenodo DOI to JOSS, please see https://github.com/openjournals/joss-reviews/issues/1622#issuecomment-517802770
