+++ README.md
@@ -133,6 +133,10 @@ Thanks\!](https://img.shields.io/badge/say-thanks-ff69b4.svg)](https://saythanks
 
 <td align="left">
 
+<a href="https://doi.org/10.21105/joss.01627">
+<img src="https://joss.theoj.org/papers/10.21105/joss.01627/status.svg" alt="DOI badge" >
+</a>
+
 </td>
 
 <td align="left">
@@ -302,7 +306,7 @@ Query and download.
 XenaQuery(df_todo) %>%
   XenaDownload() -> xe_download
 #> This will check url status, please be patient.
-#> All downloaded files will under directory /var/folders/mx/rfkl27z90c96wbmn3_kjk8c80000gn/T//Rtmp8CSlqo.
+#> All downloaded files will under directory /Volumes/Data/Rtmp/RtmpGhHoOv.
 #> The 'trans_slash' option is FALSE, keep same directory structure as Xena.
 #> Creating directories for datasets...
 #> Downloading TCGA.LUAD.sampleMap/LUAD_clinicalMatrix.gz
