+++ README.Rmd
@@ -83,7 +83,9 @@ Development
 [![Say Thanks!](https://img.shields.io/badge/say-thanks-ff69b4.svg)](https://saythanks.io/to/ShixiangWang)
 </td>
 <td align="left">
-
+<a href="https://doi.org/10.21105/joss.01627">
+  <img src="https://joss.theoj.org/papers/10.21105/joss.01627/status.svg" alt="DOI badge" >
+</a>
 </td>
 <td align="left">
 <a href="https://www.repostatus.org/#active"><img src="https://www.repostatus.org/badges/latest/active.svg" alt="Project Status: Active – The project has reached a stable, usable state and is being actively developed." /></a>
