### 2019-08-05 179c940
Fix JOSS badge showing :smile:
