+++ README.Rmd
@@ -83,8 +83,7 @@ Development
 [![Say Thanks!](https://img.shields.io/badge/say-thanks-ff69b4.svg)](https://saythanks.io/to/ShixiangWang)
 </td>
 <td align="left">
-<a href="https://doi.org/10.21105/joss.01627">
-  <img src="https://joss.theoj.org/papers/10.21105/joss.01627/status.svg" alt="DOI badge" >
+<a href="https://doi.org/10.21105/joss.01627"><img src="https://joss.theoj.org/papers/10.21105/joss.01627/status.svg" alt="JOSS" >
 </a>
 </td>
 <td align="left">
