### 2019-08-05 #315 @ShixiangWang
@melvidoni The package has published on JOSS <https://joss.theoj.org/papers/10.21105/joss.01627>. Thanks for your work, the process is really fast.
