### 2019-08-06 #315 @stefaniebutland
> can I submit the draft within a month?

Yes, sounds good. I'll mark my calendar to check with you on Aug 27. Thank you.
